+++ codegen/bin/generate-spec.js
@@ -1,2 +1,2 @@
-#!/usr/bin/env matter-run
+#!/usr/bin/env -S matter-run --max-old-heap-size=6144
 import "../dist/esm/generate-spec.js";