+++ packages/model/test/MatterTest.ts
@@ -22,7 +22,7 @@ describe("MatterDefinition", () => {
 
     it("has not increased in errors", () => {
         validate().report();
-        expect(validationResult?.errors.length).most(16);
+        expect(validationResult?.errors.length).most(2);
     });
 
     it("has not decreased in scope", () => {