+++ packages/protocol/src/interaction/InteractionServer.ts
@@ -1078,7 +1078,11 @@ export class InteractionServer implements ProtocolHandler, InteractionRecipient 
             await subscription.close(); // Cleanup
             if (error instanceof StatusResponseError) {
                 logger.info(`Sending status response ${error.code} for interaction error: ${error.message}`);
-                await messenger.sendStatus(error.code);
+                await messenger.sendStatus(error.code, {
+                    logContext: {
+                        for: "I/SubscriptionSeed-Status",
+                    },
+                });
             }
             await messenger.close();
             return; // Make sure to not bubble up the exception
@@ -1098,6 +1102,12 @@ export class InteractionServer implements ProtocolHandler, InteractionRecipient 
                 maxInterval,
                 interactionModelRevision: Specification.INTERACTION_MODEL_REVISION,
             }),
+            {
+                logContext: {
+                    subId: subscriptionId,
+                    maxInterval,
+                },
+            },
         );
 
         // When an error occurs while sending the response, the subscription is not yet active and will be cleaned up by GC
@@ -1236,12 +1246,21 @@ export class InteractionServer implements ProtocolHandler, InteractionRecipient 
                             `Send ${lastMessageProcessed ? "final " : ""}invoke response for ${invokeResponseMessage.invokeResponses} commands`,
                         );
                     }
+                    const moreChunkedMessages = lastMessageProcessed ? undefined : true;
                     await messenger.send(
                         MessageType.InvokeResponse,
                         TlvInvokeResponseForSend.encode({
                             ...invokeResponseMessage,
-                            moreChunkedMessages: lastMessageProcessed ? undefined : true,
+                            moreChunkedMessages,
                         }),
+                        {
+                            logContext: {
+                                invokeMsgFlags: Diagnostic.asFlags({
+                                    suppressResponse,
+                                    moreChunkedMessages,
+                                }),
+                            },
+                        },
                     );
                     invokeResponseMessage.invokeResponses = [];
                     messageSize = emptyInvokeResponseBytes.length;