+++ packages/protocol/src/interaction/ServerSubscription.ts
@@ -898,6 +898,7 @@ export class ServerSubscription extends Subscription {
                         interactionModelRevision: Specification.INTERACTION_MODEL_REVISION,
                     },
                     this.criteria.isFabricFiltered,
+                    !this.isClosed, // Do not wait for ack when closed
                 );
             } else {
                 await messenger.sendDataReport(
@@ -930,6 +931,7 @@ export class ServerSubscription extends Subscription {
                         }),
                     },
                     this.criteria.isFabricFiltered,
+                    !this.isClosed, // Do not wait for ack when closed
                 );
             }
         } catch (error) {